requirements.txt
Django==2.2.3
django-rest-framework==0.1.0
djangorestframework==3.9.4
pytz==2019.1
sqlparse==0.3.0
python-decouple==3.1
psycopg2-binary==2.8.3
dj-database-url==0.5.0
gunicorn==19.9.0
whitenoise==4.1.3
